StdioServer

Server that uses standard input and output streams for message transport.

For an equivalent class implemented in Node.js see Executa's StdioServer.

Super classes

rasta::Server -> rasta::StreamServer -> StdioServer

Public fields

address

The address of the server

Methods

Public methods

StdioServer$new()

StdioServer$addresses()

StdioServer$clone()

Inherited methods

rasta::Server$check()

rasta::Server$receive()

rasta::Server$start()

rasta::Server$url()

rasta::StreamServer$read()

rasta::StreamServer$stop()

rasta::StreamServer$write()

Method new()

Initialize the server.

Usage

StdioServer$new(executor = NULL)

Arguments

executor

The executor to serve

Method addresses()

Get a list of server addresses Override of Server$addresses.

Usage

StdioServer$addresses()

Method clone()

The objects of this class are cloneable with this method.

Usage

StdioServer$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.